+++ WSS_Main/src/PatternCalibModule/ParametersStructure.h
@@ -12,41 +12,41 @@
 struct Aopt_Kopt_ThreadArgs
 {
 	int 			port;
-	double 			freq;
+	float 			freq;
 
-	double 			result_Aopt;
-	double 			result_Kopt;
+	float 			result_Aopt;
+	float 			result_Kopt;
 };
 
 struct Aatt_Katt_ThreadArgs
 {
 	int 			port;
-	double 			freq;
-	double 			ATT;
+	float 			freq;
+	float 			ATT;
 
-	double 			result_Aatt;
-	double 			result_Katt;
+	float 			result_Aatt;
+	float 			result_Katt;
 };
 
 struct Sigma_ThreadArgs
 {
 	int 			port;
-	double 			freq;
-	double 			temperature;
+	float 			freq;
+	float 			temperature;
 
-	double 			result_Sigma;
+	float 			result_Sigma;
 };
 
 struct Pixel_Pos_ThreadArgs
 {
-	double 			f1;
-	double 			f2;
-	double 			fc;
-	double 			temperature;
+	float 			f1;
+	float 			f2;
+	float 			fc;
+	float 			temperature;
 
-	double 			result_F1_PixelPos;
-	double 			result_F2_PixelPos;
-	double 			result_FC_PixelPos;
+	float 			result_F1_PixelPos;
+	float 			result_F2_PixelPos;
+	float 			result_FC_PixelPos;
 };
 
 #endif /* SRC_PATTERNCALIBMODULE_PARAMETERSSTRUCTURE_H_ */
